test: keep trace classification audit-clean
